requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai
google-adk
gunicorn
Pillow
